Add schema version and command type headers to raft log
Version the log entries that raft stores, to help us in the future if we
need to change the format. In addition, add a command type to the
header, preparing for allowing callers to remove entries, and for when
emission is triggered by a raft command apply.

Only one byte is used for each field, but this should be more than
enough for us.